warnings .size() eliminados

+++ client_parserProtocoloCliente.cpp
@@ -28,7 +28,7 @@ ParserProtocoloCliente::ParserProtocoloCliente(Morph &morph, std::string &cad) :
 	std::string objSlotPreview;
 
 	std::cout << "Cadena: " << cad << std::endl;
-	while (pCad < cad.size()) {
+	while (pCad < (int)cad.size()) {
 		slotName = getCampo();
 		pCad++; //Salimos del caracter separador
 
@@ -75,7 +75,7 @@ ParserProtocoloCliente::ParserProtocoloCliente(Morph &morph, std::string &cad) :
 
 std::string ParserProtocoloCliente::getCampo() {
 	std::string campo;
-	while (cad[pCad] != CHAR_SEPARADOR and pCad < cad.size()) {
+	while (cad[pCad] != CHAR_SEPARADOR and pCad < (int)cad.size()) {
 		campo += cad[pCad];
 		pCad++;
 	}
